+++ web/src/services/theme/emotion.d.ts
@@ -1,5 +1,5 @@
 import "@emotion/react";
-import { Theme as ReearthTheme } from "./types";
+import { Theme as ReearthTheme } from "./reearthTheme/types";
 
 declare module "@emotion/react" {
   // eslint-disable-next-line @typescript-eslint/no-empty-interface
