+++ web/src/services/theme/publishTheme/types.ts
@@ -19,3 +19,15 @@ export type Colors = {
     background: string;
   };
 };
+
+export type PublishTheme = {
+  strongText: string;
+  mainText: string;
+  weakText: string;
+  strongIcon: string;
+  mainIcon: string;
+  weakIcon: string;
+  select: string;
+  mask: string;
+  background: string;
+};
