+++ web/src/services/theme/reearthTheme/lightTheme/index.ts
@@ -1,10 +1,11 @@
 import commonTheme from "../common";
 import commonColors from "../common/colors";
 import darkColors from "../darkTheme/colors";
+import type { Theme } from "../types";
 
 import colors from "./colors";
 
-const lightTheme = {
+const lightTheme: Theme = {
   ...commonTheme,
   colors: {
     ...commonColors,
