--- web/src/services/theme/types/index.ts
@@ -1,1 +1,0 @@
-export type { Theme, PublishTheme } from "./theme";
